# 8274276: Cache normalizedBase URL in URLClassPath.FileLoader